requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
pandas>=2.2.0
numpy>=1.26.0
PyYAML>=6.0.1
python-dotenv>=1.0.0

# ── Technical Indicators ──────────────────────────────────────────────────────
pandas-ta>=0.3.14b

# ── Backtest ──────────────────────────────────────────────────────────────────
vectorbt>=0.26.0

# ── Telegram ──────────────────────────────────────────────────────────────────
python-telegram-bot>=21.0.0
httpx>=0.27.0

# ── MetaTrader5 (Windows only — comment out on macOS/Linux dev) ───────────────
# MetaTrader5>=5.0.45

# ── Async / Concurrency ───────────────────────────────────────────────────────
aiofiles>=23.2.1

# ── Plotting / Reports ────────────────────────────────────────────────────────
plotly>=5.20.0
kaleido>=0.2.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.1.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
